requirements.txt
asgiref==3.8.1
Django==5.1.6
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
